UCP/GTEST: support the Rendezvous protocol and test it in gtest.
based on Yossi's initial implementation.
currently rendezvous is done by the receive side with get_zcopy.
the default rendezvous threshold is set to 1m.
to change this value from the command line, set the UCX_RNDV_THRESH
environment parameter.

solves #493.

+++ src/ucp/proto/proto.h
@@ -21,7 +21,7 @@
  */
 typedef struct {
     uint64_t                  sender_uuid;
-    uint64_t                  reqptr;
+    uintptr_t                 reqptr;
 } UCS_S_PACKED ucp_request_hdr_t;
 
 
@@ -57,7 +57,7 @@ ucs_status_t ucp_proto_progress_am_bcopy_single(uct_pending_req_t *self);
 
 /*
  * Make sure the remote worker would be able to send replies to our endpoint.
- * Should be used before a sending a message which requires a reply.
+ * Should be used before sending a message which requires a reply.
  */
 static inline void ucp_ep_connect_remote(ucp_ep_h ep)
 {
